# web/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "UiTdatabank v3",
        "description": "Version 3 of the UiTdatabank, a central database of cultural offers in the Flanders region.\n\nMost operations require that you authenticate first with your UiTID.",
        "version": "3.0.0",
        "contact": {
            "name": "CultuurNet Vlaanderen vzw",
            "url": "http://www.cultuurnet.be",
            "email": "info@uitdatabank.be"
        }
    },
    "basePath": "/",
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json",
        "application/ld+json",
        "application/xml"
    ],
    "paths": {
...
        "/event/{cdbid}/{lang}/title": {
            "post": {
                "summary": "Translate the title of a cultural event.",
                "description": "Authentication is required.",
                "operationId": "translateEventTitle",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "cdbid",
                        "in": "path",
                        "description": "The unique identifier of the cultural event.",
                        "required": true,
                        "type": "string",
                        "format": "uuid"
                    },
                    {
                        "name": "lang",
                        "in": "path",
                        "description": "The language of the translation.",
                        "required": true,
                        "type": "string",
                        "enum": [
                            "fr",
                            "de",
                            "en"
                        ]
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "An object containing the title in the given language for the given cultural event.",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/postObjectWithEventTitle"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Info on the command that was queued.",
                        "schema": {
                            "$ref": "#/definitions/command"
                        }
                    }
                },
                "tags": [
                    "Event"
                ]
            }
        },
...
        "/event/{cdbid}/labels": {
            "post": {
                "summary": "Add a label to a single cultural event.",
                "description": "Authentication is required.",
                "operationId": "addLabelToEvent",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "cdbid",
                        "in": "path",
                        "description": "The unique identifier of the cultural event.",
                        "required": true,
                        "type": "string",
                        "format": "uuid"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "An object containing the label to attach to the cultural event.",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/labelCommand"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Info on the command that was queued.",
                        "schema": {
                            "$ref": "#/definitions/command"
                        }
                    }
                },
                "tags": [
                    "Event",
                    "Label"
                ]
            }
        },
        "/events/label": {
            "post": {
                "summary": "Add a label to multiple cultural events.",
                "description": "Authentication is required",
                "operationId": "addLabelToEvents",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "An object containing information about the label and the cultural events.",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/postObjectWithlabelAndMultipleEvents"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Info on the command that was queued.",
                        "schema": {
                            "$ref": "#/definitions/command"
                        }
                    }
                },
                "tags": [
                    "Event",
                    "Label"
                ]
            }
        },
...
        "/event/{cdbid}/labels/{label}": {
            "delete": {
                "summary": "Remove a label from a cultural event.",
                "description": "Authentication is required.",
                "operationId": "removeLabelFromEvent",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "cdbid",
                        "in": "path",
                        "description": "The unique identifier of the cultural event.",
                        "required": true,
                        "type": "string",
                        "format": "uuid"
                    },
                    {
                        "name": "label",
                        "in": "path",
                        "description": "The label to remove from the cultural event.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Info on the command that was queued.",
                        "schema": {
                            "$ref": "#/definitions/command"
                        }
                    }
                },
                "tags": [
                    "Event",
                    "Label"
                ]
            }
        },
...
        "/query/label": {
            "post": {
                "summary": "Add a label to a query",
                "description": "Authentication is required.",
                "operationId": "addLabelToQuery",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "An object containing information about the label and the query.",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/postObjectWithLabelAndQuery"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Info on the command that was queued.",
                        "schema": {
                            "$ref": "#/definitions/command"
                        }
                    }
                },
                "tags": [
                    "Query",
                    "Label"
                ]
            }
        },
        "/places": {
            "get": {
                "summary": "List places.",
                "description": "",
                "operationId": "listPlaces",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "zipcode",
                        "in": "query",
                        "description": "Limit the list to places located in the area with this postal code.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of places matching the supplied criteria.",
                        "schema": {
                            "$ref": "#/definitions/PlacesList"
                        }
                    }
                },
                "tags": [
                    "Place"
                ]
            }
        },
...
        "/place/{cdbid}/labels": {
            "post": {
                "summary": "Add a label to a single place.",
                "description": "Authentication is required.",
                "operationId": "addLabelToPlace",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "cdbid",
                        "in": "path",
                        "description": "The unique identifier of the place.",
                        "required": true,
                        "type": "string",
                        "format": "uuid"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "An object containing the label to attach to the place.",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/labelCommand"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Info on the command that was queued.",
                        "schema": {
                            "$ref": "#/definitions/command"
                        }
                    }
                },
                "tags": [
                    "Place",
                    "Label"
                ]
            }
        },
        "/place/{cdbid}/labels/{label}": {
            "delete": {
                "summary": "Remove a label from a place.",
                "description": "Authentication is required.",
                "operationId": "removeLabelFromPlace",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "cdbid",
                        "in": "path",
                        "description": "The unique identifier of the place.",
                        "required": true,
                        "type": "string",
                        "format": "uuid"
                    },
                    {
                        "name": "label",
                        "in": "path",
                        "description": "The label to remove from the place.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Info on the command that was queued.",
                        "schema": {
                            "$ref": "#/definitions/command"
                        }
                    }
                },
                "tags": [
                    "Place",
                    "Label"
                ]
            }
        },
        "/place/{cdbid}/{lang}/title": {
            "post": {
                "summary": "Translate the title of a place.",
                "description": "Authentication is required.",
                "operationId": "translatePlaceTitle",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "cdbid",
                        "in": "path",
                        "description": "The unique identifier of the place.",
                        "required": true,
                        "type": "string",
                        "format": "uuid"
                    },
                    {
                        "name": "lang",
                        "in": "path",
                        "description": "The language of the translation.",
                        "required": true,
                        "type": "string",
                        "enum": [
                            "fr",
                            "de",
                            "en"
                        ]
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "An object containing the title in the given language for the given place.",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/postObjectWithPlaceTitle"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Info on the command that was queued.",
                        "schema": {
                            "$ref": "#/definitions/command"
                        }
                    }
                },
                "tags": [
                    "Place"
                ]
            }
        },
        "/place/{cdbid}/{lang}/description": {
            "post": {
                "summary": "Translate the description of a place.",
                "description": "Authentication is required.",
                "operationId": "translatePlaceDescription",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "cdbid",
                        "in": "path",
                        "description": "The unique identifier of the place.",
                        "required": true,
                        "type": "string",
                        "format": "uuid"
                    },
                    {
                        "name": "lang",
                        "in": "path",
                        "description": "The language of the translation.",
                        "required": true,
                        "type": "string",
                        "enum": [
                            "fr",
                            "de",
                            "en"
                        ]
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "An object containing the description in the given language for the given place.",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/postObjectWithPlaceDescription"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Info on the command that was queued.",
                        "schema": {
                            "$ref": "#/definitions/command"
                        }
                    }
                },
                "tags": [
                    "Place"
                ]
            }
        },
...
    },
...
    "definitions": {
...
        "command": {
            "required": [
                "commandId"
            ],
            "properties": {
                "commandId": {
                    "type": "string",
                    "example": "8cdc13e62efaecb9d8c21d59a29b9de4"
                }
            }
        },
...
        "postObjectWithEventTitle": {
            "properties": {
                "title": {
                    "type": "string",
                    "example": "An event title"
                }
            },
            "required": [
                "title"
            ]
        },
        "labelCommand": {
            "required": [
                "label"
            ],
            "properties": {
                "label": {
                    "type": "string",
                    "example": "Bio"
                }
            }
        },
        "postObjectWithPlaceDescription": {
            "properties": {
                "description": {
                    "type": "string",
                    "example": "This is a description of a place."
                }
            },
            "required": [
                "description"
            ]
        },
        "postObjectWithPlaceTitle": {
            "properties": {
                "title": {
                    "type": "string",
                    "example": "A place title"
                }
            },
            "required": [
                "title"
            ]
        },
        "postObjectWithlabelAndMultipleEvents": {
            "properties": {
                "label": {
                    "type": "string",
                    "example": "Bio"
                },
                "events": {
                    "type": "array",
                    "items": {
                        "type": "string",
                        "format": "uuid",
                        "example": "6e169e4a-32c8-4fca-a1ce-5beb2b2ac2cc"
                    }
                }
            },
            "required": [
                "label",
                "events"
            ]
        },
        "postObjectWithLabelAndQuery": {
            "properties": {
                "label": {
                    "type": "string",
                    "example": "Bio"
                },
                "query": {
                    "type": "string",
                    "example": "6e169e4a-32c8-4fca-a1ce-5beb2b2ac2cc"
                }
            },
            "required": [
                "label",
                "query"
            ]
        },
...
        "PlacesList": {
            "allOf": [
                {
                    "$ref": "#/definitions/PagedCollection"
                },
                {
                    "properties": {
                        "member": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Place"
                            }
                        }
                    }
                }
            ]
        },
...
    }
}